Neural Network Method
Summary:
The red line in the scatter plot usually represents a line of best fit through the actual popularity scores, serving as a benchmark to compare against the model's predicted scores
The blue dots represent the predicted scores.
The green line represents the line of best fit through the predicted values.